Campaigns: PHPUnit deprecation warnings.
Replacing calls to assertContains() with assertStringContainsString() because of
the following warning from PHPUnit:
```
Drupal\Tests\localgov_campaigns\Functional\CampaignBlocksTest::testCampaignNavigationBlock Using assertContains() with string haystacks is deprecated and will not be supported in PHPUnit 9. Refactor your test to use assertStringContainsString() or assertStringContainsStringIgnoringCase() instead.
Using assertNotContains() with string haystacks is deprecated and will not be supported in PHPUnit 9. Refactor your test to use assertStringNotContainsString() or assertStringNotContainsStringIgnoringCase() instead.
```