All themes / sites to customize menu display.
Use case lambeth adds a field on overview to set when and if the
menu is displayed in the header horizontally, an addition and
alternative to the rhs.

By passing the overview and the page itself this just allows full
flexibity without trying to second guess what variations there are.

Localgov could add in some default options if some become more common.