Fix PHP error on subsite page when overview is unpublished
Fix #118

Set the $subsite_id and $overview_entity to NULL if the overview is
unpublished and the user doesn't have access to it.
(It's not found in loadAndAccessCheckEntitysForTreeNodes).
This also required a change in the subsite navigation twig to check that the
link has a url before attempting to output it, a simmilar change will be
needed on any theme that has implemented the template.

Note: The test fails here due to the way the blocks cache.